+++ packages/temporal-polyfill/src/internal/errorMessages.ts
@@ -26,7 +26,8 @@ export const forbiddenField = (fieldName: string) =>
   `Invalid field ${fieldName}`
 export const duplicateFields = (fieldName: string) =>
   `Duplicate field ${fieldName}`
-export const noValidFields = 'No valid fields'
+export const noValidFields = (validFields: string[]) =>
+  'No valid fields: ' + validFields.join()
 export const invalidBag = 'Invalid bag'
 
 // Class-related